Move tests for request generation into integration test
