+++ phpunit.xml.dist
@@ -11,6 +11,7 @@
   <testsuites>
     <testsuite name="all">
       <directory suffix="Test.php">tests/Unit/</directory>
+      <directory suffix="Test.php">tests/Integration/</directory>
     </testsuite>
   </testsuites>
   <groups>
