+++ src/Redmine/Psr18Client.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Redmine;
 
 use Exception;
