+++ tests/Unit/Psr18ClientTest.php
@@ -225,79 +225,4 @@ class Psr18ClientTest extends TestCase
 
         $client->getApi('do_not_exist');
     }
-
-    /**
-     * @covers \Redmine\Psr18Client
-     * @test
-     *
-     * @dataProvider createdGetRequestsData
-     */
-    public function testCreateRequestsContainsRelevantData($path, $expectedOutput)
-    {
-        $response = $this->createMock(ResponseInterface::class);
-
-        $httpClient = $this->createMock(HttpClient::class);
-        $httpClient->method('sendRequest')->will(
-            $this->returnCallback(function($request) use ($response, $expectedOutput) {
-                $content = $request->getBody()->__toString();
-
-                $cookieHeader = '';
-                $cookies = [];
-
-                foreach ($request->getCookieParams() as $k => $v) {
-                    $cookies[] = $k.'='.$v;
-                }
-
-                $headers = '';
-
-                foreach ($request->getHeaders() as $k => $v) {
-                    $headers .= $k . ": " . $request->getHeaderLine($k)."\r\n";
-                }
-
-                if (!empty($cookies)) {
-                    $cookieHeader = 'Cookie: '.implode('; ', $cookies)."\r\n";
-                }
-
-                $fullRequest = sprintf(
-                    '%s %s %s',
-                    $request->getMethod(),
-                    $request->getUri()->__toString(),
-                    $request->getProtocolVersion()
-                )."\r\n".
-                    $headers.
-                    $cookieHeader."\r\n".
-                    $content
-                ;
-
-                $this->assertSame($expectedOutput, $fullRequest);
-
-                return $response;
-            })
-        );
-
-        $requestFactory = $this->createMock(ServerRequestFactoryInterface::class);
-        $requestFactory->method('createServerRequest')->will(
-            $this->returnCallback(function($method, $uri) {
-                return new ServerRequest($method, $uri);
-            })
-        );
-
-        $client = new Psr18Client(
-            $httpClient,
-            $requestFactory,
-            $this->createMock(StreamFactoryInterface::class),
-            'http://test.local',
-            'access_token'
-        );
-
-        $client->requestGet($path);
-    }
-
-    public function createdGetRequestsData()
-    {
-        return [
-            ['/path', "GET http://test.local/path 1.1\r\nHost: test.local\r\n\r\n"],
-            ['/path.json', "GET http://test.local/path.json 1.1\r\nHost: test.local\r\n\r\n"],
-        ];
-    }
 }
